Add files via upload
This is an 8-year USA accident dashboard. 
Data was obtained from Kaggle and with Power BI

CLEANING PROCESS
The date and time of the accidents was separated into different column to understand the period with a high rate of accident
A new measure was created out of accident start time and end time to obtain the average time taken for an accident to occur
Null values in severity rate were replaced with the median value of the severity column. 
 The data type was checked and adjusted accordingly
The Data quality was checked and adjusted accordingly  

VISUALIZATION 
The data was Visualized in Power BI

FINDINGS 
With 7.73M accidents in 8 years, the USA has an average of about 966,250 accidents in a year which is 80,521 accidents in a month and 2,684 accidents in a day. 

At 6,156,981, the severity of 2 (moderate impact) had the highest Severity rate accounting for 79% of the total accident, and was 9,039.60% higher than 1, which had the lowest Count of Severity at 67,366.﻿

There is little effect of weather on the rate of accidents as the highest number of accidents occurred in fair weather and average weather is 61.66 F which is normal weather range.

With a windspeed of 7mph which means a light breeze, there is no effect of wind on the rate of accidents.  

Humidity affects the rate of accidents as average humidity is 64.8% which is above average and slightly high. 

Accidents happen in split seconds as the average accident time is 0.03 secs. 

2,850,000 accidents occurred in crossings, junctions, traffic signals, railways and stations and that is about 36.8% accidents. 

Miami followed by Houston has the highest accident rate and more accident happens on a weekday, especially on Friday. 

2022 has the highest number of accidents with 22.8% of total accidents occurring that year. 

The higher the number of people out of their houses, the more accidents occur hence, 69.2% of the total accidents occur in the day as against 30.8% of accidents that occur at night.

There is a continuous increase in accident rate from 3 am to 7 am which is the peak accident time.  There is a sharp decrease at 9 am and 10 am and another continuous  till 4pm. It is safe to say that accident peak time is divided into two parts; 6-7am and 4-5pm. 

LIMITATION 
Lack of a population density column which would have helped to determine if population has an effect on the rate of accidents.
Incomplete 2023 data which would have helped to determine if accident rate is increasing or decreasing 
No income rate/car ownership population rate which would have helped to determine if high car ownership population resulted in more accidents.